functions/source/verify-iam-database-authentication/requirements.txt
#install latest
mysql-connector-python
